+++ kernel/arch/x64/cpu_local.rs
@@ -32,7 +32,7 @@ macro_rules! __cpu_local_impl {
                     let cpu_local_base = &__cpu_local as *const _ as usize;
                     let offset = (self as *const _ as usize) - cpu_local_base;
                     let gsbase = x86::bits64::segmentation::rdgsbase() as usize;
-                    $crate::arch::x64::VAddr::new((gsbase + offset) as u64)
+                    $crate::arch::x64::VAddr::new((gsbase + offset) as usize)
                 }
             }
         }
